src/requirements.txt
alembic==1.11.2
attrs==23.1.0
blinker==1.6.2
cffi==1.15.1
click==8.1.6
cryptography==41.0.5
flasgger==0.9.7.1
Flask==3.0.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mistune==3.0.1
packaging==23.1
psycopg2==2.9.9
pycparser==2.21
PyMySQL==1.1.0
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.9.2
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==3.0.1
